Factored Xcodeproj extensions out of IOSHelper
