+++ lib/branch_io_cli/helper.rb
@@ -3,3 +3,4 @@ require "branch_io_cli/helper/methods"
 require "branch_io_cli/helper/patch_helper"
 require "branch_io_cli/helper/report_helper"
 require "branch_io_cli/helper/util"
+require "branch_io_cli/helper/xcodeproj_ext"
